添加了python package调用功能，为此修改了data agent的代码架构，现在支持api调用和命令行运行
