+++ __main__.py
@@ -1,0 +1,6 @@
+from .graph import main
+
+if __name__ == "__main__":
+    raise SystemExit(main())
+
+
