+++ graph.py
@@ -1,372 +1,7 @@
-import os
-import base64
-import io
 import argparse
 import sys
-from dotenv import load_dotenv 
-from langchain_deepseek import ChatDeepSeek
-from langgraph.prebuilt import create_react_agent
-from langchain_core.tools import tool
-from pydantic import BaseModel, Field
-import matplotlib
-import json
-import matplotlib.pyplot as plt
-import seaborn as sns
-import pandas as pd
-import pymysql
-from langchain_tavily import TavilySearch
 
-# 默认图片输出目录（可通过环境变量 IMAGE_OUTPUT_DIR 或命令行 --images_dir 覆盖）
-IMAGE_OUTPUT_DIR = os.getenv("IMAGE_OUTPUT_DIR") or os.path.join(os.getcwd(), "images")
-
-# 加载环境变量
-load_dotenv(override=True)
-
-# ✅ 创建Tavily搜索工具
-search_tool = TavilySearch(max_results=5, topic="general")
-
-# ✅ 创建SQL查询工具
-description = """
-当用户需要进行数据库查询工作时，请调用该函数。
-该函数用于在指定MySQL服务器上运行一段SQL代码，完成数据查询相关工作，
-并且当前函数是使用pymsql连接MySQL数据库。
-本函数只负责运行SQL代码并进行数据查询，若要进行数据提取，则使用另一个extract_data函数。
-"""
-
-# 定义结构化参数模型
-class SQLQuerySchema(BaseModel):
-    sql_query: str = Field(description=description)
-
-# 封装为 LangGraph 工具
-@tool(args_schema=SQLQuerySchema)
-def sql_inter(sql_query: str) -> str:
-    """
-    当用户需要进行数据库查询工作时，请调用该函数。
-    该函数用于在指定MySQL服务器上运行一段SQL代码，完成数据查询相关工作，
-    并且当前函数是使用pymsql连接MySQL数据库。
-    本函数只负责运行SQL代码并进行数据查询，若要进行数据提取，则使用另一个extract_data函数。
-    :param sql_query: 字符串形式的SQL查询语句，用于执行对MySQL中telco_db数据库中各张表进行查询，并获得各表中的各类相关信息
-    :return：sql_query在MySQL中的运行结果。   
-    """
-    # print("正在调用 sql_inter 工具运行 SQL 查询...")
-    
-    # 加载环境变量
-    load_dotenv(override=True)
-    host = os.getenv('HOST')
-    user = os.getenv('USER')
-    mysql_pw = os.getenv('MYSQL_PW')
-    db = os.getenv('DB_NAME')
-    port = os.getenv('PORT')
-    
-    # 创建连接
-    connection = pymysql.connect(
-        host=host,
-        user=user,
-        passwd=mysql_pw,
-        db=db,
-        port=int(port),
-        charset='utf8'
-    )
-    
-    try:
-        with connection.cursor() as cursor:
-            cursor.execute(sql_query)
-            results = cursor.fetchall()
-            # print("SQL 查询已成功执行，正在整理结果...")
-    finally:
-        connection.close()
-
-    # 将结果以 JSON 字符串形式返回
-    return json.dumps(results, ensure_ascii=False)
-
-# ✅ 创建数据提取工具
-# 定义结构化参数
-class ExtractQuerySchema(BaseModel):
-    sql_query: str = Field(description="用于从 MySQL 提取数据的 SQL 查询语句。")
-    df_name: str = Field(description="指定用于保存结果的 pandas 变量名称（字符串形式）。")
-
-# 注册为 Agent 工具
-@tool(args_schema=ExtractQuerySchema)
-def extract_data(sql_query: str, df_name: str) -> str:
-    """
-    用于在MySQL数据库中提取一张表到当前Python环境中，注意，本函数只负责数据表的提取，
-    并不负责数据查询，若需要在MySQL中进行数据查询，请使用sql_inter函数。
-    同时需要注意，编写外部函数的参数消息时，必须是满足json格式的字符串，
-    :param sql_query: 字符串形式的SQL查询语句，用于提取MySQL中的某张表。
-    :param df_name: 将MySQL数据库中提取的表格进行本地保存时的变量名，以字符串形式表示。
-    :return：表格读取和保存结果
-    """
-    print("正在调用 extract_data 工具运行 SQL 查询...")
-    
-    load_dotenv(override=True)
-    host = os.getenv('HOST')
-    user = os.getenv('USER')
-    mysql_pw = os.getenv('MYSQL_PW')
-    db = os.getenv('DB_NAME')
-    port = os.getenv('PORT')
-
-    # 创建数据库连接
-    connection = pymysql.connect(
-        host=host,
-        user=user,
-        passwd=mysql_pw,
-        db=db,
-        port=int(port),
-        charset='utf8'
-    )
-
-    try:
-        # 执行 SQL 并保存为全局变量
-        df = pd.read_sql(sql_query, connection)
-        globals()[df_name] = df
-        # print("数据成功提取并保存为全局变量：", df_name)
-        return f"✅ 成功创建 pandas 对象 `{df_name}`，包含从 MySQL 提取的数据。"
-    except Exception as e:
-        return f"❌ 执行失败：{e}"
-    finally:
-        connection.close()
-
-
-class ReadData(BaseModel):
-    df_name: str = Field(description="指定用于保存结果的 pandas 变量名称（字符串形式）。")
-    file_path: str | None = Field(default=None, description="服务器本地文件路径（.csv）")
-    file_url: str | None = Field(default=None, description="文件的可直连 URL（含 LangSmith Studio 拖拽上传后的预签名链接）")
-    file_b64: str | None = Field(default=None, description="CSV 原始内容的 Base64 编码字符串")
-    csv_text: str | None = Field(default=None, description="纯文本的 CSV 内容（直接粘贴）")
-    sep: str | None = Field(default=None, description="分隔符，留空时自动嗅探。常见为 ',' 或 '\t'")
-    encoding: str | None = Field(default=None, description="文件编码，留空则由 pandas 猜测或使用系统默认")
-@tool(args_schema=ReadData)
-def read_data(
-    df_name: str,
-    file_path: str | None = None,
-    file_url: str | None = None,
-    file_b64: str | None = None,
-    csv_text: str | None = None,
-    sep: str | None = None,
-    encoding: str | None = None,
-):
-    """
-    用于读取 CSV 数据并保存为 pandas DataFrame。
-    支持以下来源（按优先级）：
-    1. file_path: 服务器本地路径
-    2. file_url: 直连 URL（包含 LangSmith Studio 拖拽上传后的预签名链接）
-    3. file_b64: Base64 编码的 CSV 字节内容
-    4. csv_text: 纯文本 CSV 内容
-    
-    :param df_name: DataFrame 在会话中的变量名
-    :return：表格读取与保存结果
-    """
-    print("正在调用 read_data 工具读取文件...")
-
-    try:
-        # pandas 自动嗅探分隔符需要 engine='python' 且 sep=None
-        read_kwargs = {}
-        if sep is None:
-            read_kwargs["sep"] = None
-            read_kwargs["engine"] = "python"
-        else:
-            read_kwargs["sep"] = sep
-        if encoding is not None:
-            read_kwargs["encoding"] = encoding
-
-        if file_path:
-            df = pd.read_csv(file_path, **read_kwargs)
-        elif file_url:
-            df = pd.read_csv(file_url, **read_kwargs)
-        elif file_b64:
-            try:
-                raw = base64.b64decode(file_b64)
-            except Exception as _:
-                return "❌ Base64 内容无法解码，请确认编码是否正确。"
-            df = pd.read_csv(io.BytesIO(raw), **read_kwargs)
-        elif csv_text:
-            df = pd.read_csv(io.StringIO(csv_text), **read_kwargs)
-        else:
-            return "⚠️ 未提供文件来源。请在 file_path、file_url、file_b64 或 csv_text 中至少提供一个。"
-
-        globals()[df_name] = df
-        return f"✅ 成功创建 pandas 对象 `{df_name}`，包含从 CSV 读取的数据，形状为 {df.shape}。"
-    except Exception as e:
-        return f"❌ 执行失败：{e}"
-
-
-# ✅创建Python代码执行工具
-# Python代码执行工具结构化参数说明
-class PythonCodeInput(BaseModel):
-    py_code: str = Field(description="一段合法的 Python 代码字符串，例如 '2 + 2' 或 'x = 3\\ny = x * 2'")
-
-@tool(args_schema=PythonCodeInput)
-def python_inter(py_code):
-    """
-    当用户需要编写Python程序并执行时，请调用该函数。
-    该函数可以执行一段Python代码并返回最终结果，需要注意，本函数只能执行非绘图类的代码，若是绘图相关代码，则需要调用fig_inter函数运行。
-    """    
-    g = globals()
-    try:
-        # 尝试如果是表达式，则返回表达式运行结果
-        return str(eval(py_code, g))
-    # 若报错，则先测试是否是对相同变量重复赋值
-    except Exception as e:
-        global_vars_before = set(g.keys())
-        try:            
-            exec(py_code, g)
-        except Exception as e:
-            return f"代码执行时报错{e}"
-        global_vars_after = set(g.keys())
-        new_vars = global_vars_after - global_vars_before
-        # 若存在新变量
-        if new_vars:
-            result = {var: g[var] for var in new_vars}
-            # print("代码已顺利执行，正在进行结果梳理...")
-            return str(result)
-        else:
-            # print("代码已顺利执行，正在进行结果梳理...")
-            return "已经顺利执行代码"
-
-# ✅ 创建绘图工具
-# 绘图工具结构化参数说明
-class FigCodeInput(BaseModel):
-    py_code: str = Field(description="要执行的 Python 绘图代码，必须使用 matplotlib/seaborn 创建图像并赋值给变量")
-    fname: str = Field(description="图像对象的变量名，例如 'fig'，用于从代码中提取并保存为图片")
-
-@tool(args_schema=FigCodeInput)
-def fig_inter(py_code: str, fname: str) -> str:
-    """
-    当用户需要使用 Python 进行可视化绘图任务时，请调用该函数。
-
-    注意：
-    1. 所有绘图代码必须创建一个图像对象，并将其赋值为指定变量名（例如 `fig`）。
-    2. 必须使用 `fig = plt.figure()` 或 `fig = plt.subplots()`。
-    3. 不要使用 `plt.show()`。
-    4. 请确保代码最后调用 `fig.tight_layout()`。
-    5. 所有绘图代码中，坐标轴标签（xlabel、ylabel）、标题（title）、图例（legend）等文本内容，必须使用英文描述。
-
-    示例代码：
-    fig = plt.figure(figsize=(10,6))
-    plt.plot([1,2,3], [4,5,6])
-    fig.tight_layout()
-    """
-    # print("正在调用fig_inter工具运行Python代码...")
-
-    current_backend = matplotlib.get_backend()
-    matplotlib.use('Agg')
-
-    local_vars = {"plt": plt, "pd": pd, "sns": sns}
-    
-    # ✅ 设置图像保存路径（默认使用 IMAGE_OUTPUT_DIR）
-    images_dir = IMAGE_OUTPUT_DIR
-    os.makedirs(images_dir, exist_ok=True)
-
-    try:
-        g = globals()
-        exec(py_code, g, local_vars)
-        g.update(local_vars)
-
-        fig = local_vars.get(fname, None)
-        if fig:
-            image_filename = f"{fname}.png"
-            abs_path = os.path.join(images_dir, image_filename)
-            fig.savefig(abs_path, bbox_inches='tight')
-            rel_path = os.path.relpath(abs_path, os.getcwd())
-            return f"✅ 图片已保存，路径为: {rel_path}"
-        else:
-            return "⚠️ 图像对象未找到，请确认变量名正确并为 matplotlib 图对象。"
-    except Exception as e:
-        return f"❌ 执行失败：{e}"
-    finally:
-        plt.close('all')
-        matplotlib.use(current_backend)
-
-# ✅ 创建提示词模板
-prompt = """
-你是一名经验丰富的智能数据分析助手，擅长帮助用户高效完成以下任务：
-
-1. **数据库查询：**
-   - 当用户需要获取数据库中某些数据或进行SQL查询时，请调用`sql_inter`工具，该工具已经内置了pymysql连接MySQL数据库的全部参数，包括数据库名称、用户名、密码、端口等，你只需要根据用户需求生成SQL语句即可。
-   - 你需要准确根据用户请求生成SQL语句，例如 `SELECT * FROM 表名` 或包含条件的查询。
-
-2. **数据表提取：**
-   - 当用户希望将数据库中的表格导入Python环境进行后续分析时，请调用`extract_data`工具。
-   - 你需要根据用户提供的表名或查询条件生成SQL查询语句，并将数据保存到指定的pandas变量中。
-
-3. **非绘图累任务的Python代码执行：**
-   - 当用户需要执行Python脚本或进行数据处理、统计计算时，请调用`python_inter`工具。
-   - 仅限执行非绘图类代码，例如变量定义、数据分析等。
-
-4. **绘图类Python代码执行：**
-   - 当用户需要进行可视化展示（如生成图表、绘制分布等）时，请调用`fig_inter`工具。
-   - 你可以直接读取数据并进行绘图，不需要借助`python_inter`工具读取图片。
-   - 你应根据用户需求编写绘图代码，并正确指定绘图对象变量名（如 `fig`）。
-   - 当你生成Python绘图代码时必须指明图像的名称，如fig = plt.figure()或fig = plt.subplots()创建图像对象，并赋值为fig。
-   - 不要调用plt.show()，否则图像将无法保存。
-
-5. **网络搜索：**
-   - 当用户提出与数据分析无关的问题（如最新新闻、实时信息），请调用`search_tool`工具。
-
-6. **文件读取（CSV）：**
-   - 当用户上传或提供 CSV 文件时，请调用`read_data`工具。
-   - 你可以从以下来源读取：`file_path`（本地路径）、`file_url`（如 LangSmith/LangGraph Studio 拖拽上传后生成的预签名链接）、`file_b64`（Base64 编码内容）或 `csv_text`（纯文本）。
-   - 读取后会在会话中创建名为 `df_name` 的 pandas 对象用于后续分析或绘图。
-
-**工具使用优先级：**
-- 如需数据库数据，请先使用`sql_inter`或`extract_data`获取，再执行Python分析或绘图。
-- 如需绘图，请先确保数据已加载为pandas对象。
- - 如用户提供了 CSV，优先使用`read_data`加载为 DataFrame。
-
-**回答要求：**
-- 所有回答均使用**简体中文**，清晰、礼貌、简洁。
-- 如果调用工具返回结构化JSON数据，你应提取其中的关键信息简要说明，并展示主要结果。
-- 若需要用户提供更多信息，请主动提出明确的问题。
-- 如果有生成的图片文件，请务必在回答中使用Markdown格式插入图片，如：![Categorical Features vs Churn](images/fig.png)
-- 不要仅输出图片路径文字。
-
-**风格：**
-- 专业、简洁、以数据驱动。
-- 不要编造不存在的工具或数据。
-
-请根据以上原则为用户提供精准、高效的协助。
-"""
-
-# ✅ 创建工具列表
-tools = [search_tool, python_inter, fig_inter, sql_inter, extract_data, read_data]
-
-# ✅ 创建模型
-model = ChatDeepSeek(model="deepseek-chat")
-
-# ✅ 创建图 （Agent）
-graph = create_react_agent(model=model, tools=tools, prompt=prompt)
-
-
-def _content_to_text(content) -> str:
-    if isinstance(content, str):
-        return content
-    if isinstance(content, list):
-        parts = []
-        for chunk in content:
-            if isinstance(chunk, dict):
-                if "text" in chunk:
-                    parts.append(str(chunk.get("text")))
-                elif "content" in chunk:
-                    parts.append(str(chunk.get("content")))
-                else:
-                    parts.append(str(chunk))
-            else:
-                parts.append(str(chunk))
-        return "\n".join(parts)
-    return str(content)
-
-
-def _print_last_ai_message(state) -> None:
-    messages = state.get("messages", [])
-    if not messages:
-        print("<no messages>")
-        return
-    last = messages[-1]
-    # AIMessage/AnyMessage from langchain_core
-    content = getattr(last, "content", None)
-    if content is None and isinstance(last, dict):
-        content = last.get("content")
-    print(_content_to_text(content))
+from .agent import create_session, chat_once
 
 
 def main(argv: list[str] | None = None) -> int:
@@ -375,7 +10,7 @@ def main(argv: list[str] | None = None) -> int:
     parser.add_argument("--interactive", action="store_true", help="Force interactive mode even if --prompt is provided.")
     parser.add_argument("--df_name", type=str, default="df", help="DataFrame variable name to store loaded CSV.")
     parser.add_argument("--csv_path", type=str, default=None, help="Local CSV file path to preload.")
-    parser.add_argument("--csv_url", type=str, default=None, help="Direct URL of CSV (e.g., LangSmith/Studio presigned URL).")
+    parser.add_argument("--csv_url", type=str, default=None, help="Direct URL of CSV (e.g., presigned).")
     parser.add_argument("--csv_b64", type=str, default=None, help="Base64-encoded CSV bytes to preload.")
     parser.add_argument("--csv_text", type=str, default=None, help="Raw CSV text to preload.")
     parser.add_argument("--sep", type=str, default=None, help="CSV delimiter (default: auto-detect).")
@@ -384,29 +19,21 @@ def main(argv: list[str] | None = None) -> int:
 
     args = parser.parse_args(argv)
 
-    # Override image output directory if provided
-    global IMAGE_OUTPUT_DIR
-    if args.images_dir:
-        IMAGE_OUTPUT_DIR = args.images_dir
-
-    # Optionally preload CSV into a DataFrame
-    if any([args.csv_path, args.csv_url, args.csv_b64, args.csv_text]):
-        # Use tool's invoke API per langchain-core >=0.1.47
-        result = read_data.invoke({
-            "df_name": args.df_name,
-            "file_path": args.csv_path,
-            "file_url": args.csv_url,
-            "file_b64": args.csv_b64,
-            "csv_text": args.csv_text,
-            "sep": args.sep,
-            "encoding": args.encoding,
-        })
-        print(str(result))
-
-    # Interactive mode if requested or no single prompt provided
+    # interactive session
     if args.interactive or not args.prompt:
         print("Interactive chat mode. Type 'exit' or 'quit' to leave.")
-        history = []
+        session = create_session(images_dir=args.images_dir)
+        if any([args.csv_path, args.csv_url, args.csv_b64, args.csv_text]):
+            preload_msg = session.preload_csv(
+                df_name=args.df_name,
+                file_path=args.csv_path,
+                file_url=args.csv_url,
+                file_b64=args.csv_b64,
+                csv_text=args.csv_text,
+                sep=args.sep,
+                encoding=args.encoding,
+            )
+            print(str(preload_msg))
         while True:
             try:
                 user_input = input("You > ").strip()
@@ -416,15 +43,24 @@ def main(argv: list[str] | None = None) -> int:
             if user_input.lower() in {"exit", "quit"}:
                 print("Bye.")
                 return 0
-            state = graph.invoke({"messages": (history + [("human", user_input)]) if history else [("human", user_input)]})
-            _print_last_ai_message(state)
-            history = state.get("messages", history)
-        # unreachable
+            print(session.send(user_input))
     else:
-        state = graph.invoke({"messages": [("human", args.prompt)]})
-        _print_last_ai_message(state)
+        preload = None
+        if any([args.csv_path, args.csv_url, args.csv_b64, args.csv_text]):
+            preload = {
+                "df_name": args.df_name,
+                "file_path": args.csv_path,
+                "file_url": args.csv_url,
+                "file_b64": args.csv_b64,
+                "csv_text": args.csv_text,
+                "sep": args.sep,
+                "encoding": args.encoding,
+            }
+        print(chat_once(args.prompt, images_dir=args.images_dir, preload=preload))
         return 0
 
 
 if __name__ == "__main__":
     sys.exit(main())
+
+
