+++ langgraph.json
@@ -1,7 +1,7 @@
 {
     "dependencies": ["./"],
     "graphs": {
-        "data_agent": "./graph.py:graph"
+        "data_agent": "./agent.py:graph"
     },
     "env": ".env"
 }
